new component Input
