+++ src/components/Task.tsx
@@ -6,6 +6,7 @@ import SizedBox from "./SizeBox"
 import check from "../assets/icons/Check.svg"
 import { Tag } from "./Tag"
 import { Row } from "./Flex"
+import Input from "./Input"
 
 interface IProps {
   align?: "row" | "column"
@@ -49,20 +50,6 @@ const Check = styled.div`
   flex-shrink: 0;
 `
 
-const TaskTitle = styled(Text)`
-  font-size: 20px;
-  font-weight: 700;
-`
-const EditableTaskTitle = styled.input`
-  font-size: 28px;
-  font-weight: 600;
-  opacity: 0.8;
-  border: none;
-  border-bottom: 1px solid #000;
-  outline: none;
-  background: transparent;
-`
-
 const Description = styled(Text)`
   font-size: 16px;
   font-weight: 400;
@@ -73,51 +60,10 @@ const StyledRow = styled(Row)`
 `
 
 const Task: React.FC<IProps> = ({ ...props }) => {
-  const [taskTitle, setTaskTitle] = useState("Enter task title")
-  const [editing, setEditing] = useState(false)
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  const handleTaskTitleClick = () => {
-    setEditing(true)
-  }
-
-  const handleTaskTitleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    setTaskTitle(event?.target.value)
-  }
-  const handleTaskTitleFix = () => {
-    setEditing(false)
-  }
-  const handleTaskTitleKeyDown = (
-    event: React.KeyboardEvent<HTMLInputElement>
-  ) => {
-    if (event.key === "Enter" || event.key === "Escape") {
-      setEditing(false)
-    }
-  }
-  useEffect(() => {
-    if (editing && inputRef.current) {
-      inputRef.current.focus()
-      const length = taskTitle.length
-      inputRef.current.setSelectionRange(length, length)
-    }
-  }, [editing, taskTitle])
-
   return (
     <Root {...props}>
       <StyledRow>
-        {editing ? (
-          <EditableTaskTitle
-            ref={inputRef}
-            value={taskTitle}
-            onChange={handleTaskTitleChange}
-            onBlur={handleTaskTitleFix}
-            onKeyDown={handleTaskTitleKeyDown}
-          />
-        ) : (
-          <TaskTitle onDoubleClick={handleTaskTitleClick}>
-            {taskTitle}
-          </TaskTitle>
-        )}
+        <Input />
         <Bin className="remove-task-button" />
       </StyledRow>
       <SizedBox height={10} />
